refactor: split Codex and tracker subsystem checks out of doctor.go (#661)
Pure behavior-preserving move along the subsystem boundaries the #661
prioritization comment names: the Codex checks (CLI/version-pin drift,
auth mode, model config, sandbox self-test, app-server JSON-RPC probe
and their codex-specific helpers) move to doctor_codex.go, and the
tracker-facing checks (Linear API-key/GraphQL auth, GitHub
agent-credential preflight and the clone-URL/repo-selection helpers)
move to doctor_tracker.go, leaving doctor.go at 553 lines — under the
800-line production-file budget. Operator-facing diagnostic wording is
untouched. The file's entry is removed from the file-size baseline so
the ratchet now enforces the budget for it directly.
